requirements.txt
typing-extensions
librosa==0.9.2
numpy
pesq==0.0.3
PyYAML==6.0
scipy
soundfile==0.12.1
wandb
pytz
python-dateutil
tqdm==4.62.3
matplotlib
pystoi
